feat: [최지수] 상품 정보 삭제요청 시 DB에서 해당 정보를 삭제합니다 [BPP-7]

+++ src/main/java/com/example/demo/product/service/ProductServiceImpl.java
@@ -68,7 +68,13 @@ public class ProductServiceImpl implements  ProductService{
     }
 
     @Override
-    public void delete(Long productId){
-        productRepository.deleteById(productId);
+    public Boolean delete(Long productId){
+        Optional<Product> maybeProduct = productRepository.findById(productId);
+
+        if(maybeProduct.isPresent()){
+            productRepository.deleteById(productId);
+            return true;
+        }
+        return false;
     }
 }
